apps/discord/requirements.txt
flask-discord-interactions
python-dotenv